Updated to 2024 framework
