+++ .vscode/settings.json
@@ -1,5 +1,5 @@
 {
-  "java.configuration.updateBuildConfiguration": "interactive",
+  "java.configuration.updateBuildConfiguration": "disabled",
   "java.import.gradle.enabled": false,
   "files.exclude": {
     "**/.git": true,
@@ -16,6 +16,8 @@
   },
   "C_Cpp.default.configurationProvider": "vscode-wpilib",
   "files.associations": {
+    "iostream": "cpp",
+    "any": "cpp",
     "array": "cpp",
     "atomic": "cpp",
     "bit": "cpp",
@@ -41,6 +43,7 @@
     "cwchar": "cpp",
     "cwctype": "cpp",
     "deque": "cpp",
+    "forward_list": "cpp",
     "list": "cpp",
     "map": "cpp",
     "set": "cpp",
@@ -69,13 +72,13 @@
     "initializer_list": "cpp",
     "iomanip": "cpp",
     "iosfwd": "cpp",
-    "iostream": "cpp",
     "istream": "cpp",
     "limits": "cpp",
     "mutex": "cpp",
     "new": "cpp",
     "numbers": "cpp",
     "ostream": "cpp",
+    "ranges": "cpp",
     "semaphore": "cpp",
     "span": "cpp",
     "sstream": "cpp",
@@ -86,8 +89,7 @@
     "cinttypes": "cpp",
     "typeindex": "cpp",
     "typeinfo": "cpp",
-    "variant": "cpp",
-    "*.inc": "cpp"
-  },
-  "C_Cpp.errorSquiggles": "enabled"
+    "valarray": "cpp",
+    "variant": "cpp"
+  }
 }
