+++ .wpilib/wpilib_preferences.json
@@ -1,6 +1,6 @@
 {
     "enableCppIntellisense": true,
     "currentLanguage": "cpp",
-    "projectYear": "2023",
+    "projectYear": "2024",
     "teamNumber": 2358
 }
