--- README.md
@@ -1,7 +1,0 @@
-# Smokey-XVII
-
-## Getting Started
-
-## Contributing
-
-## Other Helpful Tips and Tricks
