--- photonvision_config/models/labels.txt
@@ -1,1 +1,0 @@
-note
