+++ src/main/include/Collector.h
@@ -1,7 +1,7 @@
 #pragma once
-#include <rev/CANSparkMax.h>
-#include <ctre/Phoenix.h>
+#include <ctre/phoenix6/TalonFX.hpp>
 #include "BeamBreak.h"
+#include "Shooter.h"
 
 class Collector {
     public:
@@ -12,10 +12,13 @@ class Collector {
         void indexToAmp();
         void stopIndexer();
         void update();
+        void setSpeed(double rpm);
+        void stopShooter();
     private:
         // rev::CANSparkMax collectorMotor;
         // rev::CANSparkMax indexerMotor;
-        TalonFX indexerMotor;
-        TalonFX collectorMotor;
-        // BeamBreak beamBreak;
+        ctre::phoenix6::hardware::TalonFX indexerMotor;
+        ctre::phoenix6::hardware::TalonFX collectorMotor;
+        BeamBreak beamBreak;
+        Shooter a_Shooter;
 };
