+++ vendordeps/photonlib.json
@@ -1,20 +1,36 @@
 {
     "fileName": "photonlib.json",
     "name": "photonlib",
-    "version": "v2023.3.0",
-    "uuid": "515fe07e-bfc6-11fa-b3de-0242ac130004 ",
+    "version": "v2024.2.6",
+    "uuid": "515fe07e-bfc6-11fa-b3de-0242ac130004",
+    "frcYear": "2024",
     "mavenUrls": [
         "https://maven.photonvision.org/repository/internal",
         "https://maven.photonvision.org/repository/snapshots"
     ],
-    "jsonUrl": "https://maven.photonvision.org/repository/internal/org/photonvision/PhotonLib-json/1.0/PhotonLib-json-1.0.json",
+    "jsonUrl": "https://maven.photonvision.org/repository/internal/org/photonvision/photonlib-json/1.0/photonlib-json-1.0.json", 
     "jniDependencies": [],
     "cppDependencies": [
         {
             "groupId": "org.photonvision",
-            "artifactId": "PhotonLib-cpp",
-            "version": "v2023.3.0",
-            "libName": "Photon",
+            "artifactId": "photonlib-cpp",
+            "version": "v2024.2.6",
+            "libName": "photonlib",
+            "headerClassifier": "headers",
+            "sharedLibrary": true,
+            "skipInvalidPlatforms": true,
+            "binaryPlatforms": [
+                "windowsx86-64",
+                "linuxathena",
+                "linuxx86-64",
+                "osxuniversal"
+            ]
+        },
+        {
+            "groupId": "org.photonvision",
+            "artifactId": "photontargeting-cpp",
+            "version": "v2024.2.6",
+            "libName": "photontargeting",
             "headerClassifier": "headers",
             "sharedLibrary": true,
             "skipInvalidPlatforms": true,
@@ -29,13 +45,13 @@
     "javaDependencies": [
         {
             "groupId": "org.photonvision",
-            "artifactId": "PhotonLib-java",
-            "version": "v2023.3.0"
+            "artifactId": "photonlib-java",
+            "version": "v2024.2.6"
         },
         {
             "groupId": "org.photonvision",
-            "artifactId": "PhotonTargeting-java",
-            "version": "v2023.3.0"
+            "artifactId": "photontargeting-java",
+            "version": "v2024.2.6"
         }
     ]
 }
